+++ projects/framework-houdini/release_notes.md
@@ -3,6 +3,7 @@
 ## v24.11.0rc1
 2024-11-19
 
+* [fix] Plugin; Framerange is now set correctly at startup.
 * [fix] Plugin; Publisher is not launched in docked mode anymore. 
 * [changed] Dependencies; upgrade ftrack dependencies.